requirements.txt
Flask~=3.0.3
validators~=0.34.0
requests~=2.32.3
python-nmap~=0.7.1